backend/requirements.txt
# Windows-Compatible Requirements for Smart_chatbot Backend
# Install with: pip install -r requirements.txt

# Core FastAPI
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# Pydantic v2 with pre-built wheels for Windows
pydantic>=2.8.0
pydantic-settings>=2.3.0

# Database
sqlalchemy>=2.0.31
psycopg2-binary>=2.9.9
alembic>=1.13.2

# Security & Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.3

# HTTP & File handling
python-multipart>=0.0.9
httpx>=0.27.0

# Environment
python-decouple>=3.8
python-dotenv>=1.0.1

# Validation
email-validator>=2.2.0

# Date utilities
python-dateutil>=2.9.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.7

# Type hints (let pip resolve the best version)
typing-extensions>=4.12.2

# Async support
anyio>=4.4.0

# HTTP client dependencies (let pip resolve these automatically)
certifi
httpcore
idna
sniffio